+++ Sources/Utils/Extensions/Foundation/Notification+Ext.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Николай on 16.09.2022.
+//
+
+import Foundation
+
+public extension Notification.Name {
+    
+    static let deviceHaveBeenShaken = Notification.Name("deviceHaveBeenShaken")
+}
